extract generic top to worker base implementation
